refactor: float to double

+++ src/main/java/io/solar/dto/CourseDto.java
@@ -12,8 +12,8 @@ public class CourseDto {
     private Long id;
     private Long objectId;
     private Long time;
-    private Float accelerationX;
-    private Float accelerationY;
+    private Double accelerationX;
+    private Double accelerationY;
     private Long nextId;
     private Instant createdAt;
     private Long expireAt;
